+++ src/main/java/cs6310/Pokemon/Mew.java
@@ -12,7 +12,7 @@ public class Mew extends Pokemon {
         setStats();
     }
 
-    public Mew(long seed) {
+    public Mew(int seed) {
         super(seed);
         setStats();
     }
